reflexion_agent/pyproject.toml
[project]
name = "reflexion_agent"
version = "0.1.0"
description = ""
authors = [
    {name = "Alvaro"}
]
readme = "README.md"
requires-python = ">=3.13,<4.0"
dependencies = [
    "python-dotenv (>=1.1.1,<2.0.0)",
    "black (>=25.1.0,<26.0.0)",
    "isort (>=6.0.1,<7.0.0)",
    "langchain (>=0.3.27,<0.4.0)",
    "langchain-google-vertexai (>=2.0.27,<3.0.0)",
    "langgraph (>=0.5.4,<0.6.0)",
    "langchain-core (>=0.3.72,<0.4.0)",
    "langchain-google-genai (>=2.1.8,<3.0.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
package-mode = false
